fix(Ask/Bid): Cambiar el tipo de objeto que devuelve value

+++ src/main/java/edu/estatuas/stockx/Item/Ask.java
@@ -17,7 +17,7 @@ public class Ask implements Offer{
 
     @Override
     public int value() {
-        return this.ask;
+        return this.ask.intValue();
     }
 
     @Override
